backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==24.3
django-storages==1.14.4
django-unfold==0.47.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
